backend/pyproject.toml
[project]
name = "backend"
version = "3.0.0"
description = "Todo AI Chatbot - Phase 3 with OpenAI SDK and OpenRouter"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "sql>=2022.4.0",
    "fastapi>=0.111.1",
    "uvicorn[standard]>=0.30.1",
    "sqlmodel>=0.0.22",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "psycopg2-binary>=2.9.9",
    "python-multipart>=0.0.9",
    # Phase 3 - AI Chatbot dependencies
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "httpx>=0.27.0",
]
